+++ providers/dns/liquidweb/servermock_test.go
@@ -14,7 +14,7 @@ import (
 	"github.com/liquidweb/liquidweb-go/types"
 )
 
-func mockAPIServer(t *testing.T, initRecs ...network.DNSRecord) string {
+func mockAPIServer(t *testing.T, initRecs []network.DNSRecord) string {
 	t.Helper()
 
 	recs := make(map[int]network.DNSRecord)
@@ -66,6 +66,7 @@ func requireJSON(next http.Handler) http.HandlerFunc {
 
 func mockAPICreate(recs map[int]network.DNSRecord) http.HandlerFunc {
 	_, mockAPIServerZones := makeMockZones()
+
 	return func(w http.ResponseWriter, r *http.Request) {
 		body, err := io.ReadAll(r.Body)
 		if err != nil {
@@ -135,6 +136,7 @@ func mockAPIDelete(recs map[int]network.DNSRecord) http.HandlerFunc {
 
 func mockAPIListZones() http.HandlerFunc {
 	mockZones, mockAPIServerZones := makeMockZones()
+
 	return func(w http.ResponseWriter, r *http.Request) {
 		body, err := io.ReadAll(r.Body)
 		if err != nil {
@@ -170,6 +172,134 @@ func mockAPIListZones() http.HandlerFunc {
 			http.Error(w, string(respBody), http.StatusOK)
 			return
 		}
+
 		http.Error(w, "", http.StatusInternalServerError)
 	}
+}
+
+func makeEncodingError(buf []byte) string {
+	return fmt.Sprintf(`{"data":"%q","encoding":"JSON","error":"unexpected end of string while parsing JSON string, at character offset 32 (before \"(end of string)\") at /usr/local/lp/libs/perl/LW/Base/Role/Serializer.pm line 16.\n","error_class":"LW::Exception::Deserialize","full_message":"Could not deserialize \"%q\" from JSON: unexpected end of string while parsing JSON string, at character offset 32 (before \"(end of string)\") at /usr/local/lp/libs/perl/LW/Base/Role/Serializer.pm line 16.\n"}⏎`, string(buf), string(buf))
+}
+
+func makeMockZones() (map[int]network.DNSZoneList, map[string]int) {
+	mockZones := map[int]network.DNSZoneList{
+		1: {
+			Items: []network.DNSZone{
+				{
+					ID:                1,
+					Name:              "blars.com",
+					Active:            1,
+					DelegationStatus:  "CORRECT",
+					PrimaryNameserver: "ns.liquidweb.com",
+				},
+				{
+					ID:                2,
+					Name:              "tacoman.com",
+					Active:            1,
+					DelegationStatus:  "CORRECT",
+					PrimaryNameserver: "ns.liquidweb.com",
+				},
+				{
+					ID:                3,
+					Name:              "storm.com",
+					Active:            1,
+					DelegationStatus:  "CORRECT",
+					PrimaryNameserver: "ns.liquidweb.com",
+				},
+				{
+					ID:                4,
+					Name:              "not-apple.com",
+					Active:            1,
+					DelegationStatus:  "BAD_NAMESERVERS",
+					PrimaryNameserver: "ns.liquidweb.com",
+				},
+				{
+					ID:                5,
+					Name:              "example.com",
+					Active:            1,
+					DelegationStatus:  "BAD_NAMESERVERS",
+					PrimaryNameserver: "ns.liquidweb.com",
+				},
+			},
+		},
+		2: {
+			Items: []network.DNSZone{
+				{
+					ID:                6,
+					Name:              "banana.com",
+					Active:            1,
+					DelegationStatus:  "NXDOMAIN",
+					PrimaryNameserver: "ns.liquidweb.com",
+				},
+				{
+					ID:                7,
+					Name:              "cherry.com",
+					Active:            1,
+					DelegationStatus:  "SERVFAIL",
+					PrimaryNameserver: "ns.liquidweb.com",
+				},
+				{
+					ID:                8,
+					Name:              "dates.com",
+					Active:            1,
+					DelegationStatus:  "SERVFAIL",
+					PrimaryNameserver: "ns.liquidweb.com",
+				},
+				{
+					ID:                9,
+					Name:              "eggplant.com",
+					Active:            1,
+					DelegationStatus:  "SERVFAIL",
+					PrimaryNameserver: "ns.liquidweb.com",
+				},
+				{
+					ID:                10,
+					Name:              "fig.com",
+					Active:            1,
+					DelegationStatus:  "UNKNOWN",
+					PrimaryNameserver: "ns.liquidweb.com",
+				},
+			},
+		},
+		3: {
+			Items: []network.DNSZone{
+				{
+					ID:                11,
+					Name:              "grapes.com",
+					Active:            1,
+					DelegationStatus:  "UNKNOWN",
+					PrimaryNameserver: "ns.liquidweb.com",
+				},
+				{
+					ID:                12,
+					Name:              "money.banana.com",
+					Active:            1,
+					DelegationStatus:  "UNKNOWN",
+					PrimaryNameserver: "ns.liquidweb.com",
+				},
+				{
+					ID:                13,
+					Name:              "money.stand.banana.com",
+					Active:            1,
+					DelegationStatus:  "UNKNOWN",
+					PrimaryNameserver: "ns.liquidweb.com",
+				},
+				{
+					ID:                14,
+					Name:              "stand.banana.com",
+					Active:            1,
+					DelegationStatus:  "UNKNOWN",
+					PrimaryNameserver: "ns.liquidweb.com",
+				},
+			},
+		},
+	}
+
+	mockAPIServerZones := make(map[string]int)
+	for _, page := range mockZones {
+		for _, zone := range page.Items {
+			mockAPIServerZones[zone.Name] = int(zone.ID)
+		}
+	}
+	return mockZones, mockAPIServerZones
 }
